All the code moved to the subpackage named "logic" - additional nesting will prevent from mixing the code of this project with github-users-web and github-users-android
