+++ src/main/java/com/xemantic/githubusers/logic/model/User.java
@@ -18,26 +18,18 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-package com.xemantic.githubusers.event;
-
-import com.xemantic.githubusers.model.User;
-
-import java.util.Objects;
+package com.xemantic.githubusers.logic.model;
 
 /**
- * Event generated when user is selected.
+ * User JSON representation according to
+ * <a href="https://developer.github.com/v3/search/#search-users">GitHub API</a>.
+ *
+ * @author morisil
  */
-public class UserSelectedEvent {
+public interface User {
 
-    private final User user;
+  String getLogin();
 
-    public UserSelectedEvent(User user) {
-        this.user = Objects.requireNonNull(user);
-    }
-
-    public User getUser() {
-        return user;
-    }
+  String getAvatarUrl();
 
 }
