+++ src/main/java/com/xemantic/githubusers/logic/view/UserView.java
@@ -18,19 +18,21 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.xemantic.githubusers.service;
 
-import com.xemantic.githubusers.model.SearchResult;
-import rx.Single;
+package com.xemantic.githubusers.logic.view;
+
+import com.xemantic.githubusers.logic.event.Trigger;
+import rx.Observable;
 
 /**
- * User search service according to
- * <a href="https://developer.github.com/v3/search/#search-users">GitHub API</a>.
+ * Displays single user entry.
  *
  * @author morisil
  */
-public interface UserService {
+public interface UserView {
 
-  Single<SearchResult> find(String query, int page, int perPage);
+  void displayLogin(String login);
+
+  Observable<Trigger> observeSelection();
 
 }
