+++ src/test/java/com/xemantic/githubusers/logic/presenter/UserPresenterTest.java
@@ -19,14 +19,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.xemantic.githubusers.presenter;
+package com.xemantic.githubusers.logic.presenter;
 
-import com.xemantic.githubusers.event.Trigger;
-import com.xemantic.githubusers.event.UserSelectedEvent;
-import com.xemantic.githubusers.eventbus.EventBus;
-import com.xemantic.githubusers.eventbus.EventTracker;
-import com.xemantic.githubusers.model.User;
-import com.xemantic.githubusers.view.UserView;
+import com.xemantic.githubusers.logic.event.Trigger;
+import com.xemantic.githubusers.logic.event.UserSelectedEvent;
+import com.xemantic.githubusers.logic.eventbus.EventBus;
+import com.xemantic.githubusers.logic.eventbus.EventTracker;
+import com.xemantic.githubusers.logic.model.User;
+import com.xemantic.githubusers.logic.view.UserView;
 import org.junit.Test;
 import rx.Observable;
 import rx.subjects.PublishSubject;
